fix: 修复 CI 编译错误和警告
修复的问题：
1. resource_browser.cpp 缺少 <functional> 头文件导致编译失败
2. 修复类型转换警告（-Wconversion）
   - Segment::progress() 和 SegmentStats::progress() 中的除法转换
   - calculate_adaptive_segments() 中的整数到浮点转换
3. 修复符号转换警告（-Wsign-conversion）
   - resource_browser.cpp 中的 depth 参数
4. 标记未使用参数以消除警告
   - task_manager.cpp 中的 priority 和 task_id 参数
   - segment_downloader.cpp 中的辅助函数使用 [[maybe_unused]]

测试结果：
- 所有 191 个测试通过
- 编译成功，仅剩少量无害警告

+++ packages/libfalcon/include/falcon/segment_downloader.hpp
@@ -44,7 +44,7 @@ struct Segment {
     /// Get progress ratio (0.0 ~ 1.0)
     [[nodiscard]] float progress() const noexcept {
         Bytes seg_size = size();
-        return seg_size > 0 ? static_cast<float>(downloaded.load()) / seg_size : 1.0f;
+        return seg_size > 0 ? static_cast<float>(downloaded.load()) / static_cast<float>(seg_size) : 1.0f;
     }
 };
 
@@ -61,7 +61,7 @@ struct SegmentStats {
     /// Get overall progress (0.0 ~ 1.0)
     [[nodiscard]] float progress() const noexcept {
         Bytes total = total_size.load();
-        return total > 0 ? static_cast<float>(total_downloaded.load()) / total : 0.0f;
+        return total > 0 ? static_cast<float>(total_downloaded.load()) / static_cast<float>(total) : 0.0f;
     }
 
     /// Get download speed in bytes/second
